+++ src/main/java/com/sbproject/piclocket/security/SecurityConfig.java
@@ -1,0 +1,42 @@
+package com.sbproject.piclocket.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+/**
+ * Configures API security.
+ */
+@Configuration
+public class SecurityConfig {
+
+    private final JwtAuthenticationFilter filter;
+
+    public SecurityConfig(JwtAuthenticationFilter filter) {
+        this.filter = filter;
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        return http
+                .csrf(AbstractHttpConfigurer::disable)
+                .sessionManagement(
+                        session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                )
+                .authorizeHttpRequests(
+                        auth -> auth.requestMatchers(
+                                "/actuator/health",
+                                "/dev/token"
+                        ).permitAll().anyRequest().authenticated()
+                )
+                .addFilterBefore(
+                        filter,
+                        UsernamePasswordAuthenticationFilter.class
+                )
+                .build();
+    }
+}